requirements.txt
accelerate==0.13.1
numpy==1.23.3
pandas==1.5.0
python_Levenshtein==0.20.5
scikit_learn==1.1.2
scipy==1.9.1
sentence_transformers==2.2.2
seqeval==1.2.2
sklearn_pandas==2.2.0
tabulate==0.9.0
torch==1.12.1+cu113
tqdm==4.64.1
transformers==4.22.2
undecorated==0.3.0
datasets==2.5.2